backend/requirements.txt
# ========================================
# SmartDocAI Backend Dependencies
# ========================================

# ---- Core Web ----
fastapi>=0.103,<0.120
uvicorn>=0.23,<0.31
python-dotenv>=1.0
python-multipart>=0.0.7

# ---- DB ----
sqlalchemy>=2.0,<2.1
pydantic>=2.4,<3
pydantic-settings>=2.0,<3
psycopg2-binary>=2.9

# ---- Email Validation (Required by Pydantic EmailStr) ----
email-validator>=2.0

# ---- Auth ----
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.0,<4.2

# ---- Document Processing (CRITICAL PINS) ----
pdfplumber>=0.10,<0.12
pdfminer.six==20221105
python-docx>=1.0
pytesseract>=0.3
pdf2image>=1.16
Pillow>=9.5,<11

# ---- ML CORE ----
# NOTE: PyTorch CPU-only versions installed via Dockerfile
# This saves ~2GB by excluding CUDA/GPU libraries:
#   torch==2.0.1+cpu (installed in Dockerfile)
#   torchvision==0.15.2+cpu (installed in Dockerfile)
#   torchaudio==2.0.2+cpu (installed in Dockerfile)

# ---- Embeddings ----
sentence-transformers>=2.2,<3.4
faiss-cpu>=1.7,<1.8

# ---- NLP ----
transformers>=4.30,<4.39
tokenizers>=0.13,<0.16
safetensors>=0.3
sentencepiece>=0.1.99

# ---- Utils ----
numpy>=1.23,<1.26
scipy>=1.9,<1.12
scikit-learn>=1.2,<1.4
requests>=2.30
tqdm>=4.64
regex>=2023.6
duckduckgo-search>=5.0

# ---- API Enhancements ----
slowapi>=0.1.9